add dependencies.py and manager script
